feat(core): establish compositional vault architecture and multi-slot liquid containers
- Pivoted the core architectural paradigm from a sheer transport protocol to a definitive ECS Container and Capability Framework, solidifying the library as the ultimate logistical storage endpoint.
- Architected the `Vault` infrastructure to centrally manage Container Maps and Blueprints, seamlessly integrating it alongside a new block routing system via strict object composition within the terminal `Component<ChunkStore>`.
- Engineered the `SimpleLiquidContainer` to support multi-slot fluid storage, underpinned by a novel `BluePrint` pattern that cleanly segregates data-driven JSON configurations from dynamic runtime container states.
- Optimized internal data structures by migrating container payload storage from generic lists to high-performance `Short2ObjectMap` allocations, significantly improving lookup overhead and memory efficiency.
- Implemented a robust reactive event pipeline via `IFluxTransaction` and `onContentsChanged` listeners, guaranteeing deterministic block/chunk data persistence upon any container content mutation.

+++ src/main/java/io/github/popo70023/fluxvault/payload/liquid/interaction/ui/LiquidContainerWindow.java
@@ -1,5 +1,5 @@
 /*
- * FluxVault - A universal transport protocol for Hytale.
+ * FluxVault - The Ultimate ECS Resource Storage & Capability Framework for Hytale.
  * Copyright (c) 2026 Ben (popo70023)
  * Licensed under the MIT License.
  */
@@ -34,7 +34,7 @@ public class LiquidContainerWindow extends BlockWindow {
         this.windowData.addProperty("blockItemId", item.getId());
         long capacity = fluxContainers.getCapacity();
         this.windowData.addProperty("containerCapacity", capacity);
-        for (int i = 0; i < fluxContainers.getContainerMaxSize(); i++) {
+        for (short i = 0; i < fluxContainers.getContainerMaxSize(); i++) {
             LiquidStack content = fluxContainers.getContent(i);
             if (content != null && !content.isEmpty()) {
                 long quantity = content.getQuantity();
